+++ bw2data/backends/proxies.py
@@ -97,6 +97,26 @@ class Activity(ActivityProxyBase):
     def id(self):
         return self._document.id
 
+    def __getitem__(self, key):
+        if key == 0:
+            return self["database"]
+        elif key == 1:
+            return self["code"]
+        elif key in self._data:
+            return self._data[key]
+
+        try:
+            rp = self.rp_exchange()
+        except ValueError:
+            raise KeyError
+
+        if key in rp.get('classifications', []):
+            return rp['classifications'][key]
+        if key in rp.get('properties', []):
+            return rp['properties'][key]
+
+        raise KeyError
+
     def __setitem__(self, key, value):
         if key == "id":
             raise ValueError("`id` is read-only")
@@ -256,6 +276,22 @@ class Activity(ActivityProxyBase):
             if include_substitution
             else ("production",),
         )
+
+    def rp_exchange(self):
+        """Return an ``Exchange`` object corresponding to the reference production. Uses the following in order:
+
+        * The ``production`` exchange, if only one is present
+        * The ``production`` exchange with the same name as the activity ``reference product``.
+
+        Raises ``ValueError`` if no suitable exchange is found."""
+        candidates = list(self.production())
+        if len(candidates) == 1:
+            return candidates[0]
+        candidates2 = [exc for exc in candidates if exc.input['name'] == self.get('reference product')]
+        if len(candidates2) == 1:
+            return candidates2[0]
+        else:
+            raise ValueError("Can't find a single reference product exchange (found {} candidates)".format(len(candidates)))
 
     def producers(self):
         return self.production()